BaseTools/edksetup.sh: Handle the return value from grep
When SetupPython3() parses the possible python binary paths, it uses
`grep "[[:digit:]]$"` to filter the python versions in the file name.
Since grep would return 1 when the last character of the file name is
not a digit, OvmfPkg/build.sh would fail immediately when the function
is handling some cases, e.g. "python3:".

This commit adds `|| true` to the grep command to make sure the command
never returns 1 in any condition.

[NOTE: For the python3 branch]

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Gary Lin <glin@suse.com>
Cc: Liming Gao <liming.gao@intel.com>
Cc: Yonghong Zhu <yonghong.zhu@intel.com>